requirements.txt
selenium==4.22.0
pytest==8.2.2
allure-pytest==2.13.5
requests==2.31.0
jsonschema==4.22.0
Appium-Python-Client==2.9.0
Pillow==10.3.0
numpy==2.0.0
pyautogui==0.9.54
PyYAML==6.0
pytest-allure-adaptor==1.7.11
behave==1.2.7
webdriver-manager==4.0.1
pyscreenshot==3.1
allure-behave==2.13.5
